cricket-intelligence/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
httpx
sse-starlette
requests
langsmith
